Implemented VCU ECU Syncronization
The algorithm will try its best so Sync the two log tables. Please note that in some cases (It seems when logs starts after some time the car is turned on) sync might be impossible since the 'on' time si unknown. If a constant delay ratio is found a backpropagation sync will be implemented.